requirements.txt
mysql-connector-python
hydra-core
IPython
hydra-joblib-launcher
torch
numpy
black
